requirements.txt
appnope==0.1.0
backports.shutil-get-terminal-size==1.0.0
decorator==4.3.0
Django==1.11.14
djangorestframework==3.8.2
enum34==1.1.6
ipython==5.8.0
ipython-genutils==0.2.0
pathlib2==2.3.2
pexpect==4.6.0
pickleshare==0.7.4
prompt-toolkit==1.0.15
ptyprocess==0.6.0
Pygments==2.2.0
pytz==2018.5
scandir==1.9.0
simplegeneric==0.8.1
six==1.11.0
traitlets==4.3.2
wcwidth==0.1.7
